renaming threat_stream.py --> main.py to be more in line with other lambda funcs
